requirements.txt
termcolor==2.4.0
python-dotenv==1.0.1
openai
groq
requests